enable configuration via file (preferring environment)
Systems like kubernetes allow you to specify configuration via a
configmap/secret which can be updated and will update the file within
the container.  Environment variables do not have this feature.  When it
comes to managing this inside a container in a kubernetes cluster, this
will allow you to update the configuration without restarting the
container.

Current configurations will not be broken as the environment variable is
still in use and preferred.  To use this feature, you must only set the
LETSENCRYPT_AWS_CONFIG_FILE environment variable but will default to
config.json in the directory its run from.