allow multiple ELBs per cert request/update
The current datastructure makes it impossible to issue a single cert to
use across multiple ELBs.  It should be possible to allow for a single
cert with many SANs across many ELBs.  This refactor makes it possible
while maintaining backwards compatibility with the old scheme (issues a
warning to users about the new schema).